A lot of changes in this version. This version fixes bugs reported by KEREVAL. The bugs fixed are for Goil: event mask generation, the ISR list (an ISR does not need to be present know, this bug fix is also for the OS code), use of OS configuration attributes (PRETASKHOOK and so on). For the operating system: a wrong test in the libpcl implementation, some macros that was wrong for extended error checking, a lock that was improperly done in GetEvent, missing error checking in alarm services, a test that was not done for basic tasks, a big mess in the management of ready tasks (activation instances was not taken into account in the good order), a wrong rescheduling when TerminateTask was in error, a missing test in ReleaseResource.

+++ os/tpl_os_custom_types.h
@@ -75,7 +75,7 @@ typedef u8 tpl_event_mask;
  *
  * Default aliased type is "unsigned short int"
  */
-typedef u16 tpl_tick;
+typedef u32 tpl_tick;
 
 /**
  * tpl_alarm_type is used for alarm identifiers.
